issue #1 (refactor - models) : switch filename chapter_dao and topic_dao
