+++ src/dao/postgres_impl/postgres_chapters_dao.ts
@@ -1,6 +1,6 @@
 import { Knex } from "knex";
 import db from "../../database/database";
-import { IChapterDAO } from "../topic_dao";
+import { IChapterDAO } from "../chapter_dao";
 import { IChapter } from "../../models/chapter";
 
 export class PostgresChapterDAO implements IChapterDAO {
